feat(remoter): 优化工具初始化时机，每次对话都重新初始化client及其工具
